Remove refresh of ../fulcro-rad-datomic
With the upcoming changes add Datomic Cloud support to
`fulcro-rad-datomic`, this line will require users to have dependencies
for Datomic Cloud and on-prem on ones classpath. We could optionally
set up a deps alias that is meant to be used if you are developing
`fulcro-rad-datomic` with the demo, but this doesn't seem like something
that should necessarily be included in the public demo, so I'm proposing
removing it.

I removed the equivalent for `../fulcro-rad`, as it seems we should
either support both or none in this repo.