Demo: Back browses through history of report controls
Normally `control/set-parameter` does `replace-route` so history of past controls states is not kept. Thus we hack the pikcer's `:action` to store the current state (by doubling it so that only the newly added copy is overwritten by set-parameter).

Then we add-route-listener! to catch cases when the URL changes but the route does not (as dr/change-route-relative! ignores those) and when the url also contains params relevant to any on-screen report and in this case we force the routing to happen - this refreshing the report.